requirements.txt
Django==1.11.11
django-filter==1.1.0
djangorestframework==3.7.7
lxml==3.4.0
Markdown==2.5.1
MySQL-python==1.2.5
pytz==2018.3
qcloudapi-sdk-python==2.0.12
simplejson==3.6.5
